app/pyproject.toml
[project]
name = "mcp_polydial"
version = "0.1.0"
description = "Minimal MCP server with no 3rd-party dependencies"
requires-python = ">=3.11,<3.13"
dependencies = [
    "mcp[cli]>=1.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "twilio>=8.10.0",
    "python-jose[cryptography]>=3.3.0",
    "email-validator>=2.0.0",
    "pytz>=2023.3",
    "pydantic>=2.0.0",
    "boto3>=1.28.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"